refactor: re-structuring of DocumentUploadDetailsService for IDV and onfido
