+++ packages/wallets/src/features/accounts/screens/DocumentUploadDetailsService/DocumentUploadDetailsService.scss
@@ -1,4 +1,6 @@
 .wallets-idv-document-details {
+    width: 100%;
+    max-width: 64.2rem;
     border-radius: 0.8rem;
     border: 0.1rem solid var(--system-light-5-active-background, #d6dadb);
     margin-bottom: 2.4rem;
