+++ packages/wallets/src/features/accounts/screens/DocumentUploadDetailsService/index.ts
@@ -1,0 +1,1 @@
+export { default as DocumentUploadDetailsService } from './DocumentUploadDetailsService';
