--- packages/wallets/src/features/accounts/screens/IDVDocumentUpload/components/IDVDocumentUploadDetails/index.ts
@@ -1,1 +1,0 @@
-export { default as IDVDocumentUploadDetails } from './IDVDocumentUploadDetails';
