--- packages/wallets/src/features/accounts/screens/IDVDocumentUpload/components/index.ts
@@ -1,1 +1,0 @@
-export * from './IDVDocumentUploadDetails';
