+++ packages/wallets/src/features/accounts/screens/index.ts
@@ -1,6 +1,7 @@
 export * from './AddressSection';
 export * from './CommonMistakesExamples';
 export * from './DocumentSubmission';
+export * from './DocumentUploadDetailsService';
 export * from './IDVDocumentUpload';
 export * from './ManualDocumentUpload';
 export * from './PersonalDetails';
