+++ packages/wallets/src/features/cfd/screens/Onfido/Onfido.scss
@@ -1,6 +1,9 @@
 .wallets-onfido {
     width: 99.6rem;
     min-height: 47.2rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
     padding: 1.6rem 0;
     overflow-y: auto;
